Add a penalty for traffic signals
Traffic signals are quite isotropic and the penalty is very small, so
results should not change much overall. The main interest of this is to
avoid following "weird" cycleways going through many traffic signals
around a junction when one could simply dismount and cross by foot in a
single step.