Add a penalty based on maximum speed of ways
Add a penalty to roads without cycleways, based on the maxspeed. This is
due to the fact that any maxspeed above 30 km/h is way faster than a
bike and not really safe so should be penalized. This would bias towards
slower and more peaceful streets.

Also take into account `highway=unknown` tags.