fix/unify the file extensions used in map creator
Creating routing data files (`*.rd5`) isn't possible at the moment:

The "node tiles" are created in `OsmCutter` with file extensions
`.ntl` [1]

In later steps the node tiles are loaded. That fails in the current
version, because different file extensions are used:

- `NodeCutter` tries to load node tiles with `.tlf` extension [2]
- `NodeFilter` tries to load node tiles with `.tls` extension [3]
- `WayCutter` tries to load node tiles with `.tlf` extension [4]

This commit changes the file extensions used when loading the
node tiles to `.ntl` in the mentioned files.

[1]: https://github.com/abrensch/brouter/blob/master/brouter-map-creator/src/main/java/btools/mapcreator/OsmCutter.java#L341
[2]: https://github.com/abrensch/brouter/blob/master/brouter-map-creator/src/main/java/btools/mapcreator/NodeCutter.java#L37
[3]: https://github.com/abrensch/brouter/blob/master/brouter-map-creator/src/main/java/btools/mapcreator/NodeFilter.java#L51
[4]: https://github.com/abrensch/brouter/blob/master/brouter-map-creator/src/main/java/btools/mapcreator/WayCutter.java#L36