fastbike & trekking: add a use_proposed_cycleroutes option
When true, consider proposed cycle routes and regular cycle routes as equals.
Default is set to false (keep current behaviour).

In some places, there are many physically present cycle routes (or just
"interesting" ways for bicycles) that are tagged as "proposed" in OSM
(sometimes because it was made for a specific event like the Olympic Games,
and then stayed permanently, sometimes because it was not made or approved by
some "official" administration).

When setting this switch to true, these will be considered like regular
cycleroutes.