Remove cycleroute always having access granted
Some bicycle routes can have portions which are not accessible.

Potential reasons for this:
 - temporary construction work on the ways (sometimes for months),
 - local access change without simultaneous update of bicycle route relation
   (lack of mapper's knowledge or lack of time to redefine an updated route),  
 - ... 

This commit changes the trekking profile so that being on a cycle route does 
not automatically mean "access granted" for bicycles. The normal access 
checkings are used instead.

However, the accesspenalty is set to 15 if the no-access way belongs to such a 
route, so that the way is not completely forbidden if there is no other 
alternative unavailable.