Improve readability of internal voice hint command's names
In discussions relating to voice hints sometimes there is still
confusion regarding the difference between names used internally in
BRouter and strings actually emitted in the output.

Also, when working with the code it is sometimes not immediately obvious
what short names like `RNLB` or `TSLR` actually stand for.

By using names unique to BRouter internally, i.e. names not used in
output formats, and making them more verbose, both issues can be
alleviated.

It is assumed `OFFR` was supposed to mean "Off route" as referred to in
the respective comment added in 23d1812371, but not "offroad" as in the
mapping for Cruiser. Note it is actually never used anywhere in the
whole codebase, so there is nothing to infer its function from.

Obviously, this will not have any impact on the API or any output
formats at all.

Test Plan:
  - No compile errors.
  - GeoJSON API output unchanged