Configure profile as service when changing profile settings
When trying to change settings for a profile that was not yet
configured as a service in the android app, the app exits instead of
configuring the profile. This behavior is not very intuitive for users.

This change instead adds a dialog to configure the profile as a new
service in case `Profile Settings` is selected and the selected profile
is not yet configured as a service.