fastbike profile: lower traffic consideration
In commit c2e6516, consider_traffic was enabled by default in fastbike. When
using this profile in big cities, this made the vast majority of main ways
very costly and so, avoided. This made big differences in routing decisions,
especially within big cities where significantly more turns can be added to
the route in trying to avoid primary/secondary/main ways that (when allowed
for bicycles) are usually no *so* bad.

This commit aims to reinstate a fastbike routing behaviour closer to the
previous experience, while considering traffic a little bit more than before.
The user is still able to change the parameter to its preference.

Also, the current code defines hascycleway but does not use this variable at
all while fastbike-verylowtraffic and mtb profiles use it to limit the traffic
penalty when a cycleway is present. I believe this was not made on purpose.

Note: Avoiding traffic closer to the spirit of the trekking profile where
consider_traffic is nevertheless disabled by default at the moment.